+++ src/main/resources/jira-sda-app.xml
@@ -20,7 +20,7 @@ http://www.springframework.org/schema/beans/spring-beans-4.0.xsd">
     </bean>
 
     <!--project-->
-    <bean id="projectControler" class="pl.sda.jira.project.ProjectControler">
+    <bean id="projectControler" class="pl.sda.jira.controler.ProjectControler">
         <constructor-arg ref="teamRepository"/>
         <constructor-arg ref="projectRepository"/>
     </bean>
