+++ src/test/java/pl/sda/jira/project/ProjectControlerTest.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import pl.sda.jira.controler.ProjectControler;
 import pl.sda.jira.project.model.Project;
 import pl.sda.jira.project.model.ProjectRepository;
 import pl.sda.jira.project.model.Team;
